# bpo-47009: Let PRECALL_NO_KW_LIST_APPEND do its own POP_TOP