@markshannon 2022-04-05
Looks good. I'm a bit wary of specialized superinstructions, but this seems solid.
I can imagine cases where `list.append()` wouldn't be followed by a `POP_TOP`, but they are contrived and highly unlikely.