@tiran 2022-04-05
The assert is failing on s390x Fedora buildbot https://buildbot.python.org/all/#/builders/232/builds/524

```
_bootstrap_python: Python/ceval.c:5045: _PyEval_EvalFrameDefault: Assertion `next_instr[-1] == POP_TOP' failed.
make: *** [Makefile:1204: Python/frozen_modules/io.h] Aborted (core dumped)
```