@markshannon 2022-04-05
Strange. The bytecode is exactly the same on all platforms.